containerの記述をコンポーネントファイルに移した
